Added collision detection for walls
Collision detection now works for all four directions and prevents the player from being able to leave the arena. This was done by adding four additional colliders as children to the player that test for triggers. Each has a bool and is stored in an array inside the player movement script. This also required adding a rigidbody2D to the wall objects.